requirements.txt
apipkg==1.4
argh==0.26.2
Babel==2.3.4
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
debtcollector==1.8.0
docopt==0.6.2
enum34==1.1.6
execnet==1.4.1
frozendict==1.0
funcsigs==1.0.2
futures==3.0.5
futurist==0.18.0
glob2==0.4.1
iso8601==0.1.11
Mako==1.0.4
MarkupSafe==0.23
monotonic==1.2
netaddr==0.7.18
netifaces==0.10.5
notifier==1.0.3
oslo.i18n==3.9.0
oslo.utils==3.16.0
parse==1.6.6
parse-type==0.3.4
pathtools==0.1.2
pbr==1.10.0
prettytable==0.7.2
py==1.4.31
pycrypto==2.6.1
pyparsing==2.1.9
-e git+git@github.com:giupo/ServiceDiscovery.git@0e735b1ea5fcc947b2ebe036d740aeceb91a0930#egg=ServiceDiscovery
pytest==3.0.2
pytest-bdd==2.17.1
pytest-cov==2.3.1
pytest-watch==4.1.0
pytest-xdist==1.15.0
pytz==2016.6.1
PyYAML==3.12
six==1.10.0
watchdog==0.8.3
weakrefmethod==1.0.3
wrapt==1.10.8
